js ui refactor

+++ src/main/java/com/dzytsiuk/onlinestore/dao/jdbc/mapper/ProductRowMapper.java
@@ -12,7 +12,7 @@ public class ProductRowMapper implements RowMapper<Product> {
     @Override
     public Product mapRow(ResultSet resultSet, int i) throws SQLException {
         Product product = new Product();
-        product.setId(resultSet.getLong("id"));
+        product.setId(resultSet.getInt("id"));
         Timestamp date = resultSet.getTimestamp("creation_date");
         product.setCreationDate(date.toLocalDateTime());
         product.setName(resultSet.getString("name"));
